pyproject.toml
[project]
name = "short-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[standard]>=0.128.0",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.57.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
]
